Add rewards indicator on Rewards page
