+++ src/hooks/useBorrowBonkRewardsAvailability.tsx
@@ -1,0 +1,16 @@
+import { useQuery } from '@tanstack/react-query'
+
+import { fetchBorrowBonkRewardsAvailability } from '@banx/api/activity'
+
+export const useBorrowBonkRewardsAvailability = () => {
+  const { data: bonkRewardsAvailable } = useQuery(
+    ['borrowBonkRewardsAvailability'],
+    () => fetchBorrowBonkRewardsAvailability(),
+    {
+      staleTime: 20 * 1000, //? 20 sec
+      refetchInterval: 30 * 1000, //? 30 sec
+      refetchOnWindowFocus: false,
+    },
+  )
+  return bonkRewardsAvailable || false
+}
