+++ src/pages/LeaderboardPage/components/RewardsTab/RewardsTab.module.less
@@ -67,6 +67,17 @@
     }
   }
 }
+.bonkRewards {
+  background: var(--additional-green-secondary);
+  color: var(--additional-green-primary-deep);
+  padding: 0 8px;
+  border-radius: 100px;
+
+  &Off {
+    background: var(--bg-secondary);
+    color: var(--content-secondary);
+  }
+}
 .weeklyRewardsBorrowBtn {
   text-decoration: none;
   button {
